Issue #8: Users > Change and Reset Password
Added "Forgot Password?" button to login page which then prompts the user to enter in their username or email address. If a match is found, then an email is sent with a link to reset their password to something of their choosing. An email is sent to confirm that the password change has occurred.

Added "Change Password" button to user profile page which then prompts the user to enter a new password. An email is sent to confirm that the password change has occurred.

Added session based flash messages to notify the user that a certain action has occurred

Added WebUtils to programmatically get the baseUrl (we may choose just to hard code to http://anotherglass.net)

Updated Params to do some basic sanitation or the request parameters